Added capability to take hourly stream and make a daily stream including joining in daily snodas data.
